requirements.txt
boto3>=1.21.21,<2.0.0
pytest<=7.2.0
pytest-asyncio<=0.20.2
cloudpathlib[s3]<=0.10.0
pandas<=1.3.3
python-dotenv<=0.19.2